backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cat-schedule-backend"
version = "0.1.0"
description = "Invite-only course schedule and grades app backend"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "apscheduler>=3.10.4",
  "beautifulsoup4>=4.12.3",
  "cryptography>=42.0.8",
  "ddddocr>=1.5.6",
  "email-validator>=2.2.0",
  "fastapi>=0.111.0",
  "passlib>=1.7.4",
  "pillow>=10.3.0",
  "psycopg[binary]>=3.1.19",
  "pydantic-settings>=2.3.3",
  "python-multipart>=0.0.9",
  "redis>=5.0.6",
  "requests>=2.32.3",
  "rich>=13.7.1",
  "rq>=1.16.2",
  "sqlalchemy>=2.0.31",
  "typer>=0.12.3",
  "uvicorn[standard]>=0.30.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.2"]

[project.scripts]
cat-schedule = "app.cli:app"
cat-schedule-scheduler = "app.tasks.scheduler:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
